Using Value : Alternate Approach to Builder in ChangeNotifierProvider
Alternate to builder in ChangeNotifierProvider: using "value"

if context is not required, we may use ".value" approach instead of "builder method"
We shoud use ".value" approach if we are using provider on something that's part of a list or a grid because widgets are recycled by flutter, but the data that is attached to the widget changes.
When using changeNotifierProvider.value, we actually make sures that provider works even if data changes for the widget.
If we use builder function here, it'll not work correctly bcz now the provider is tied to its data and is attached/detached to/from the widget instead of changing data being attached to the same provider