Using Consumer (Alternate to Provider.of)
Consumer uses builder which uses context and instance of data, and a child

Why using Consumer?
When we use Provider.of, then the whole build method will rebuild whenever that data changes.
But with Consumer, we could only wrap the subpart of the widget tree on which we want to listen state updates